Implement snakefoot requested changes for common class for asp.net buffer
