--- src/NLog.Web/Targets/Wrappers/AspNetBufferingTargetWrapper.cs
@@ -1,99 +1,0 @@
-﻿using System.Web;
-using NLog.Targets;
-
-namespace NLog.Web.Targets.Wrappers
-{
-    /// <summary>
-    /// Buffers log events for the duration of ASP.NET request and sends them down
-    /// to the wrapped target at the end of a request.
-    /// </summary>
-    /// <seealso href="https://github.com/nlog/nlog/wiki/AspNetBufferingWrapper-target">Documentation on NLog Wiki</seealso>
-    /// <remarks>
-    /// <p>
-    /// Typically this target is used in cooperation with PostFilteringTargetWrapper
-    /// to provide verbose logging for failing requests and normal or no logging for
-    /// successful requests. We need to make the decision of the final filtering rule
-    /// to apply after all logs for a page have been generated.
-    /// </p>
-    /// <p>
-    /// To use this target, you need to add an entry in the httpModules section of
-    /// web.config:
-    /// </p>
-    /// <code lang="XML">
-    /// <![CDATA[<?xml version="1.0" ?>
-    /// <configuration>
-    ///   <system.web>
-    ///     <httpModules>
-    ///       <add name="NLog" type="NLog.Web.NLogHttpModule, NLog.Extended"/>
-    ///     </httpModules>
-    ///   </system.web>
-    /// </configuration>
-    /// ]]>
-    /// </code>
-    /// </remarks>
-    /// <example>
-    /// <p>To set up the ASP.NET Buffering target wrapper <a href="config.html">configuration file</a>, put
-    /// the following in <c>web.nlog</c> file in your web application directory (this assumes
-    /// that PostFilteringWrapper is used to provide the filtering and actual logs go to
-    /// a file).
-    /// </p>
-    /// <code lang="XML" source="examples/targets/Configuration File/ASPNetBufferingWrapper/web.nlog" />
-    /// <p>
-    /// This assumes just one target and a single rule. More configuration
-    /// options are described <a href="config.html">here</a>.
-    /// </p>
-    /// <p>
-    /// To configure the target programmatically, put the following
-    /// piece of code in your <c>Application_OnStart()</c> handler in Global.asax.cs
-    /// or some other place that gets executed at the very beginning of your code:
-    /// </p>
-    /// <code lang="C#" source="examples/targets/Configuration API/ASPNetBufferingWrapper/Global.asax.cs" />
-    /// <p>
-    /// Fully working C# project can be found in the <c>Examples/Targets/Configuration API/ASPNetBufferingWrapper</c>
-    /// directory along with usage instructions.
-    /// </p>
-    /// </example>
-    [Target("AspNetBufferingWrapper", IsWrapper = true)]
-    public class AspNetBufferingTargetWrapper : AspNetBufferingTargetWrapperBase
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AspNetBufferingTargetWrapper" /> class.
-        /// </summary>
-        public AspNetBufferingTargetWrapper()
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AspNetBufferingTargetWrapper" /> class.
-        /// </summary>
-        /// <param name="wrappedTarget">The wrapped target.</param>
-        public AspNetBufferingTargetWrapper(Target wrappedTarget)
-            : base(wrappedTarget, 100)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AspNetBufferingTargetWrapper" /> class.
-        /// </summary>
-        /// <param name="wrappedTarget">The wrapped target.</param>
-        /// <param name="bufferSize">Size of the buffer.</param>
-        public AspNetBufferingTargetWrapper(Target wrappedTarget, int bufferSize)
-            : base(wrappedTarget, bufferSize)
-        {
-        }
-
-        /// <summary>
-        /// If the request is already in progress, register for this
-        /// </summary>
-        protected override void HandleRequestAlreadyBegun()
-        {
-            if (HttpContext.Current != null)
-            {
-                // we are in the context already,
-                // it's too late for OnBeginRequest to be called,
-                // so let's just call it ourselves
-                OnBeginRequest(null, new HttpContextEventArgs(HttpContext.Current));
-            }
-        }
-    }
-}
